+++ hodor-admin/src/main/java/org/dromara/hodor/admin/controller/UserController.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.dromara.hodor.admin.core.Result;
 import org.dromara.hodor.admin.core.ResultUtil;
-import org.dromara.hodor.admin.domain.User;
+import org.dromara.hodor.admin.entity.User;
 import org.dromara.hodor.admin.service.UserService;
 import org.dromara.hodor.core.PageInfo;
 import org.springframework.web.bind.annotation.DeleteMapping;
