+++ hodor-admin/src/main/java/org/dromara/hodor/admin/service/JobGroupService.java
@@ -18,7 +18,7 @@
 package org.dromara.hodor.admin.service;
 
 import org.dromara.hodor.core.PageInfo;
-import org.dromara.hodor.admin.domain.User;
+import org.dromara.hodor.admin.entity.User;
 import org.dromara.hodor.core.entity.JobGroup;
 
 /**
