+++ hodor-admin/src/main/java/org/dromara/hodor/admin/service/PermissionService.java
@@ -17,7 +17,7 @@
 
 package org.dromara.hodor.admin.service;
 
-import org.dromara.hodor.admin.domain.User;
+import org.dromara.hodor.admin.entity.User;
 
 /**
  * PermissionService
