+++ hodor-admin/src/main/java/org/dromara/hodor/admin/service/TenantService.java
@@ -1,6 +1,6 @@
 package org.dromara.hodor.admin.service;
 
-import org.dromara.hodor.admin.domain.Tenant;
+import org.dromara.hodor.admin.entity.Tenant;
 import org.dromara.hodor.core.PageInfo;
 
 /**
