+++ hodor-admin/src/main/java/org/dromara/hodor/admin/service/impl/UserGroupsServiceImpl.java
@@ -8,7 +8,7 @@ import com.github.yulichang.wrapper.MPJLambdaWrapper;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.dromara.hodor.admin.domain.UserGroups;
+import org.dromara.hodor.admin.entity.UserGroups;
 import org.dromara.hodor.admin.mapper.UserGroupsMapper;
 import org.dromara.hodor.admin.service.UserGroupsService;
 import org.dromara.hodor.core.PageInfo;
