+++ hodor-admin/src/main/resources/mappers/TenantMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="org.dromara.hodor.admin.mapper.TenantMapper">
 
-    <resultMap type="org.dromara.hodor.admin.domain.Tenant" id="TenantMap">
+    <resultMap type="org.dromara.hodor.admin.entity.Tenant" id="TenantMap">
         <result property="id" column="id" jdbcType="INTEGER"/>
         <result property="tenantName" column="tenant_name" jdbcType="VARCHAR"/>
         <result property="corpName" column="corp_name" jdbcType="VARCHAR"/>
