+++ hodor-admin/src/main/resources/mappers/UserGroupsMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="org.dromara.hodor.admin.mapper.UserGroupsMapper">
 
-    <resultMap type="org.dromara.hodor.admin.domain.UserGroups" id="UserGroupsMap">
+    <resultMap type="org.dromara.hodor.admin.entity.UserGroups" id="UserGroupsMap">
         <result property="id" column="id" jdbcType="INTEGER"/>
         <result property="userId" column="user_id" jdbcType="INTEGER"/>
         <result property="groupId" column="group_id" jdbcType="INTEGER"/>
