+++ hodor-admin/src/main/resources/mappers/UserMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="org.dromara.hodor.admin.mapper.UserMapper">
 
-    <resultMap type="org.dromara.hodor.admin.domain.User" id="UserMap">
+    <resultMap type="org.dromara.hodor.admin.entity.User" id="UserMap">
         <result property="id" column="id" jdbcType="INTEGER"/>
         <result property="username" column="username" jdbcType="VARCHAR"/>
         <result property="password" column="password" jdbcType="VARCHAR"/>
